requirements.txt
accelerate==0.31.0
bitsandbytes==0.43.3
datasets==2.21.0
diffusers==0.30.0
huggingface_hub==0.23.4
matplotlib
numpy
packaging==24.1
Pillow==10.4.0
safetensors==0.4.5
setuptools
torch
torchvision
tqdm==4.66.5
transformers==4.44.2
wandb==0.18.1
# xformers
